Chat Page divided in components
