+++ client/src/components/Chat/MsgContainer.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export function MsgContainer() {
+  return (
+        <div className="messageContainer">
+            {
+
+            }
+        </div>
+  )
+}
