+++ client/src/components/Chat/OpenCloseBtn.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export function OpenCloseBtn({ toggleLeftBar }) {
+  return (
+    <div className="openCloseBtn" onClick={toggleLeftBar}></div>
+  )
+}
